requirements.txt
Flask==3.0.0
gunicorn==21.2.0
yfinance==0.2.36
numpy==1.23.5
pandas==2.2.1
scikit-learn==1.4.1.post1
tensorflow-cpu==2.15.0
gdown==5.1.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
flask-sqlalchemy==3.1.1
Werkzeug==3.0.1